requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
ccxt==4.0.6
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
